requirements.txt
flask
flask_restful
flask_cors
mysql.connector
sqlalchemy
pyyaml
argparse
requests
passlib
pycryptodome
discord-webhook